reshuffle inducing_variables and covariances
