+++ gpflow/covariances.py
