--- gpflow/covariances/__init__.py
@@ -1,4 +1,0 @@
-from .dispatch import Kuf
-from .dispatch import Kuu
-
-from . import kufs, kuus, mo_kufs, mo_kuus
