+++ gpflow/inducing_variables/__init__.py
@@ -1,4 +1,7 @@
-from .inducing_variables import InducingVariables, InducingPoints, Multiscale
+from .base import InducingVariables
+from .inducing_points import InducingPoints
 from .inducing_patch import InducingPatches
-from .mo_inducing_variables import (MultioutputInducingVariables, FallbackSharedIndependentInducingVariables, FallbackSeparateIndependentInducingVariables,
-                                    SharedIndependentInducingVariables, SeparateIndependentInducingVariables)
+from .multiscale import Multiscale
+from .multioutputs import (MultioutputInducingVariables,
+                           FallbackSharedIndependentInducingVariables, FallbackSeparateIndependentInducingVariables,
+                           SharedIndependentInducingVariables, SeparateIndependentInducingVariables)
