+++ gpflow/inducing_variables/base.py
