+++ gpflow/inducing_variables/multioutputs.py
@@ -12,7 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .inducing_variables import InducingVariables
+from typing import Union
+
+import tensorflow as tf
+
+from ..kernels import (MultioutputKernel, SeparateIndependent, LinearCoregionalization, SharedIndependent,
+                       IndependentLatent)
+from ..covariances import Kuf, Kuu
+from . import InducingVariables, InducingPoints
 
 
 class MultioutputInducingVariables(InducingVariables):
@@ -120,3 +127,94 @@ class SeparateIndependentInducingVariables(FallbackSeparateIndependentInducingVa
     the most efficient implementation.
     """
     pass
+
+
+@Kuf.register(InducingPoints, MultioutputKernel, object)
+def _Kuf(inducing_variable: InducingPoints, kernel: MultioutputKernel, Xnew: tf.Tensor):
+    return kernel(inducing_variable.Z, Xnew, full=True, full_output_cov=True)  # [M, P, N, P]
+
+
+@Kuf.register(SharedIndependentInducingVariables, SharedIndependent, object)
+def _Kuf(inducing_variable: SharedIndependentInducingVariables, kernel: SharedIndependent, Xnew: tf.Tensor):
+    return Kuf(inducing_variable.inducing_variable_shared, kernel.kernel, Xnew)  # [M, N]
+
+
+@Kuf.register(SeparateIndependentInducingVariables, SharedIndependent, object)
+def _Kuf(inducing_variable: SeparateIndependentInducingVariables, kernel: SharedIndependent, Xnew: tf.Tensor):
+    return tf.stack([Kuf(f, kernel.kernel, Xnew)
+                     for f in inducing_variable.inducing_variable_list], axis=0)  # [L, M, N]
+
+
+@Kuf.register(SharedIndependentInducingVariables, SeparateIndependent, object)
+def _Kuf(inducing_variable: SharedIndependentInducingVariables, kernel: SeparateIndependent, Xnew: tf.Tensor):
+    return tf.stack([Kuf(inducing_variable.inducing_variable_shared, k, Xnew) for k in kernel.kernels], axis=0)  # [L, M, N]
+
+
+@Kuf.register(SeparateIndependentInducingVariables, SeparateIndependent, object)
+def _Kuf(inducing_variable: SeparateIndependentInducingVariables, kernel: SeparateIndependent, Xnew: tf.Tensor):
+    Kufs = [Kuf(f, k, Xnew) for f, k in zip(inducing_variable.inducing_variable_list, kernel.kernels)]
+    return tf.stack(Kufs, axis=0)  # [L, M, N]
+
+
+@Kuf.register((FallbackSeparateIndependentInducingVariables, FallbackSharedIndependentInducingVariables),
+              LinearCoregionalization,
+              object)
+def _Kuf(inducing_variable: Union[SeparateIndependentInducingVariables, SharedIndependentInducingVariables],
+         kernel: LinearCoregionalization, Xnew: tf.Tensor):
+    kuf_impl = Kuf.dispatch(type(inducing_variable), SeparateIndependent, object)
+    K = tf.transpose(kuf_impl(inducing_variable, kernel, Xnew), [1, 0, 2])  # [M, L, N]
+    return K[:, :, :, None] * tf.transpose(kernel.W)[None, :, None, :]  # [M, L, N, P]
+
+
+@Kuf.register(SharedIndependentInducingVariables, LinearCoregionalization, object)
+def _Kuf(inducing_variable: SharedIndependentInducingVariables, kernel: SeparateIndependent, Xnew: tf.Tensor):
+    return tf.stack([Kuf(inducing_variable.inducing_variable_shared, k, Xnew) for k in kernel.kernels], axis=0)  # [L, M, N]
+
+
+@Kuf.register(SeparateIndependentInducingVariables, LinearCoregionalization, object)
+def _Kuf(inducing_variable, kernel, Xnew):
+    return tf.stack([Kuf(f, k, Xnew)
+                     for f, k in zip(inducing_variable.inducing_variable_list, kernel.kernels)], axis=0)  # [L, M, N]
+
+
+@Kuu.register(InducingPoints, MultioutputKernel)
+def _Kuu(inducing_variable: InducingPoints, kernel: MultioutputKernel, *, jitter=0.0):
+    Kmm = kernel(inducing_variable.Z, full=True, full_output_cov=True)  # [M, P, M, P]
+    M = tf.shape(Kmm)[0] * tf.shape(Kmm)[1]
+    jittermat = jitter * tf.reshape(tf.eye(M, dtype=Kmm.dtype), tf.shape(Kmm))
+    return Kmm + jittermat
+
+
+@Kuu.register(FallbackSharedIndependentInducingVariables, SharedIndependent)
+def _Kuu(inducing_variable: FallbackSharedIndependentInducingVariables, kernel: SharedIndependent, *, jitter=0.0):
+    Kmm = Kuu(inducing_variable.inducing_variable_shared, kernel.kernel)  # [M, M]
+    jittermat = tf.eye(len(inducing_variable), dtype=Kmm.dtype) * jitter
+    return Kmm + jittermat
+
+
+@Kuu.register(FallbackSharedIndependentInducingVariables, (SeparateIndependent, IndependentLatent))
+def _Kuu(inducing_variable: FallbackSharedIndependentInducingVariables,
+         kernel: Union[SeparateIndependent, IndependentLatent],
+         *,
+         jitter=0.0):
+    Kmm = tf.stack([Kuu(inducing_variable.inducing_variable_shared, k) for k in kernel.kernels], axis=0)  # [L, M, M]
+    jittermat = tf.eye(len(inducing_variable), dtype=Kmm.dtype)[None, :, :] * jitter
+    return Kmm + jittermat
+
+
+@Kuu.register(FallbackSeparateIndependentInducingVariables, SharedIndependent)
+def _Kuu(inducing_variable: FallbackSeparateIndependentInducingVariables, kernel: SharedIndependent, *, jitter=0.0):
+    Kmm = tf.stack([Kuu(f, kernel.kernel) for f in inducing_variable.inducing_variable_list], axis=0)  # [L, M, M]
+    jittermat = tf.eye(len(inducing_variable), dtype=Kmm.dtype)[None, :, :] * jitter
+    return Kmm + jittermat
+
+
+@Kuu.register(FallbackSeparateIndependentInducingVariables, (SeparateIndependent, LinearCoregionalization))
+def _Kuu(inducing_variable: FallbackSeparateIndependentInducingVariables,
+         kernel: Union[SeparateIndependent, LinearCoregionalization],
+         *,
+         jitter=0.0):
+    Kmms = [Kuu(f, k) for f, k in zip(inducing_variable.inducing_variable_list, kernel.kernels)]
+    Kmm = tf.stack(Kmms, axis=0)  # [L, M, M]
+    jittermat = tf.eye(len(inducing_variable), dtype=Kmm.dtype)[None, :, :] * jitter
+    return Kmm + jittermat
