+++ tests/test_multioutput.py
@@ -4,7 +4,7 @@ import scipy
 import tensorflow as tf
 
 import gpflow
-import gpflow.inducing_variables.mo_inducing_variables as mf
+import gpflow.inducing_variables.multioutputs as mf
 import gpflow.kernels.mo_kernels as mk
 from gpflow.conditionals import sample_conditional
 from gpflow.conditionals.util import fully_correlated_conditional, fully_correlated_conditional_repeat, sample_mvn
